+++ main.go
@@ -13,7 +13,7 @@ import (
 	"github.com/bugsnag/bugsnag-cli/pkg/utils"
 )
 
-var package_version = "2.9.1"
+var package_version = "2.9.2"
 
 func main() {
 	commands := options.CLI{}
@@ -168,6 +168,17 @@ func main() {
 		}
 
 		err := upload.ProcessUnityAndroid(commands, endpoint, logger)
+
+		if err != nil {
+			logger.Fatal(err.Error())
+		}
+
+	case "upload breakpad <path>":
+		if commands.ApiKey == "" {
+			logger.Fatal("missing api key, please specify using `--api-key`")
+		}
+
+		err := upload.ProcessBreakpad(commands, endpoint, logger)
 
 		if err != nil {
 			logger.Fatal(err.Error())