+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## TBD
+
+### Enhancements
+- Adds support for uploading breakpad .sym files [178](https://github.com/bugsnag/bugsnag-cli/pull/178)
+
+## 2.9.2 (2025-02-11)
+
+### Fixes
+- Get NDK version from the `source.properties` file when uploading NDK symbol files [172](https://github.com/bugsnag/bugsnag-cli/pull/172)
+
 ## 2.9.1 - (2025-01-23)
 
 ### Fixes